Create guessing-game.py
I created a program consisting of a simple guessing game where the computer will generate a random number utilizing the lower and upper limits that I've created. The objective of this game is for the user to generate their own upper and lower limit for random numbers and them guessing the random number that is situated between the upper and lower limit.